Overhauls player page data processing, adds linegraph hardcoded to display fantasy points.
