+++ frontend/src/components/player_name/LineChart.tsx
@@ -1,72 +1,40 @@
 import React, { PureComponent } from 'react';
 import { LineChart, Line, XAxis, YAxis, CartesianGrid, Tooltip, Legend, ResponsiveContainer } from 'recharts';
 
-const data = [
-  {
-    name: 'Page A',
-    uv: 4000,
-    pv: 2400,
-    amt: 2400,
-  },
-  {
-    name: 'Page B',
-    uv: 3000,
-    pv: 1398,
-    amt: 2210,
-  },
-  {
-    name: 'Page C',
-    uv: 2000,
-    pv: 9800,
-    amt: 2290,
-  },
-  {
-    name: 'Page D',
-    uv: 2780,
-    pv: 3908,
-    amt: 2000,
-  },
-  {
-    name: 'Page E',
-    uv: 1890,
-    pv: 4800,
-    amt: 2181,
-  },
-  {
-    name: 'Page F',
-    uv: 2390,
-    pv: 3800,
-    amt: 2500,
-  },
-  {
-    name: 'Page G',
-    uv: 3490,
-    pv: 4300,
-    amt: 2100,
-  },
-];
-
-export default function RenderLineChart({value}) {
+export default function RenderLineChart({data, rows, cols}) {
 
   return (
     <ResponsiveContainer width="100%" height="100%">
       <LineChart
-        width={500}
-        height={300}
-        data={data}
-        margin={{
-          top: 5,
-          right: 30,
-          left: 20,
-          bottom: 5,
-        }}
+          width={500}
+          height={300}
+          data={rows}
+          margin={{
+            top: 50,
+            right: 20,
+            left: 20,
+            bottom: 20,
+          }}
       >
-        <CartesianGrid strokeDasharray="3 3" />
-        <XAxis dataKey="name" />
-        <YAxis />
-        <Tooltip />
-        <Legend />
-        <Line type="monotone" dataKey="pv" stroke="#8884d8" activeDot={{ r: 8 }} />
+        <CartesianGrid strokeDasharray="3 3"/>
+        <XAxis dataKey="week" label={{
+          value: "Game Week",
+          position: "insideBottom",
+          offset: -5,
+          style: {marginTop: 8}
+        }}/>
+        <YAxis dataKey={"fantasy_points"} type="number" domain={[0, 30]} allowDataOverflow label={{
+          value: "Fantasy Points",
+          angle: -90,
+          position: 'insideLeft',
+          style: {textAnchor: 'middle'}
+        }} ticks={[0, 5, 10, 15, 20, 25, 30]}
+        />
+        <Tooltip/>
+        <Line type="monotone" dataKey="fantasy_points" labelHidden stroke="#8884d8" activeDot={{r: 8}}/>
+        <text x={"50%"} y={"5%"} textAnchor="middle" dominantBaseline="middle" fill="#fff" fontSize={16} fontWeight="500">
+          Fantasy Points by Game Week
+        </text>
       </LineChart>
     </ResponsiveContainer>
   );
